src/api/requirements.txt
asgiref==3.9.1
attrs==25.3.0
Django==4.2.24
django-cors-headers==4.3.1
django-filter==25.1
djangorestframework==3.16.1
drf-spectacular==0.28.0
gunicorn==23.0.0
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
Pillow==11.0.0
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.27.1
sqlparse==0.5.3
typing_extensions==4.15.0
uritemplate==4.2.0
